Abstract columns into separate components
