+++ src/components/amazon-ads/search-terms/SearchTermColumn.jsx
@@ -1,0 +1,34 @@
+import { Group, Box, Text } from "@mantine/core"
+import toUpper from 'lodash-es/toUpper'
+import { TARGET_STATES } from '@/utils/constants'
+import styles from '@/styles/RecordTable.module.css'
+
+function SearchTermColumn({ row, negativeKeywords, negativeTargets }) {
+  const negativeKeyword = negativeKeywords.find(k => (
+    k.keywordText === row.searchTerm && 
+    k.adGroupId === row.adGroupId
+  ))
+
+  const negativeTarget = negativeTargets.find(t => {
+    return (
+      toUpper(t.expression?.[0]?.value) === toUpper(row.searchTerm) && 
+      t.adGroupId === row.adGroupId
+    )
+  })
+
+  return (
+    <Group gap="xs" wrap="nowrap">
+      {negativeKeyword?.state === TARGET_STATES.ENABLED && (
+        <Box className={styles.negative} title="Negative Keyword">N</Box>
+      )}
+
+      {negativeTarget?.state === TARGET_STATES.ENABLED && (
+        <Box className={styles.negative} title="Negative Target">N</Box>
+      )}
+
+      <Text size="xs">{row.searchTerm}</Text>
+    </Group>
+  )
+}
+
+export default SearchTermColumn
