+++ src/components/amazon-ads/search-terms/index.js
@@ -1,0 +1,2 @@
+export { default as KeywordColumn } from './KeywordColumn'
+export { default as SearchTermColumn } from './SearchTermColumn'
