+++ src/styles/RecordTable.module.css
@@ -57,11 +57,11 @@
   background-color: var(--mantine-color-red-6);
 }
 
-.negativeKeyword {
+.negative {
     display: flex;
     align-items: center;
     justify-content: center;
-    font-weight: 600;
+    font-weight: bold;
     font-size: 0.8rem;
     padding: 0.2rem;
     height: 1.1rem;
