+++ src/views/ads/AdGroup.jsx
@@ -4,7 +4,7 @@ import { Title, Stack, SimpleGrid, Loader } from '@mantine/core'
 
 import api from '@/api'
 import { useAsync } from '@/hooks/useAsync'
-import { KeywordRows } from '@/components/amazon_ads/Keyword'
+import { KeywordRows } from '@/components/amazon-ads/Keyword'
 
 function AdGroup() {
   const { adGroupId } = useParams()
