+++ src/views/ads/Keyword.jsx
@@ -5,7 +5,7 @@ import NotFound from '@/views/NotFound'
 import { useAsync } from '@/hooks/useAsync'
 import api from '@/api'
 import DataList from '@/components/DataList'
-import { KeywordStateSelect } from '@/components/amazon_ads/Keyword'
+import { KeywordStateSelect } from '@/components/amazon-ads/Keyword'
 import NegativeKeywordButton from '@/components/NegativeKeywordButton'
 import { TARGET_STATES } from '@/utils/constants'
