+++ src/views/ads/SearchTerms.jsx
@@ -1,9 +1,8 @@
 import { useState, useEffect, useMemo, useCallback } from "react"
-import { Stack, Title, Group, Loader, Button, Text, Box } from "@mantine/core"
+import { Stack, Title, Group, Loader, Button } from "@mantine/core"
 import { useForm } from '@mantine/form'
 import { subDays, format } from 'date-fns'
 import { useNavigate } from 'react-router-dom'
-import lowerCase from 'lodash-es/lowerCase'
 
 import api from "@/api"
 import { validators } from '@/utils/validation'
@@ -18,7 +17,7 @@ import DateRangeInputPicker from "@/components/DateRangeInputPicker"
 import { columnTypes, createDefaultSort, createDefaultFilter, getSortableColumns } from '@/utils/table'
 import { SEARCH_TERMS_HIDDEN_COLUMNS, TARGET_STATES } from '@/utils/constants'
 import { getEntityType } from '@/utils/amazon-ads'
-import styles from '@/styles/RecordTable.module.css'
+import { KeywordColumn, SearchTermColumn } from '@/components/amazon-ads/search-terms'
 
 
 const formatDate = (date) => format(date, 'yyyy-MM-dd')
@@ -178,38 +177,19 @@ function SearchTerms() {
   }, [navigate])
 
   const columnComponents = useMemo(() => ({
-    keyword: (row) => {
-      const state = keywords[row.keywordId]?.state || targets[row.keywordId]?.state
-
-      const negativeKeyword = negativeKeywords.find(k => (
-        k.keywordText === row.searchTerm && 
-        k.adGroupId === row.adGroupId
-      ))?.state
-
-      const negativeTarget = negativeTargets.find(t => (
-        t.expression?.[0]?.value === row.searchTerm && 
-        t.adGroupId === row.adGroupId
-      ))?.state
-
-      return (
-        <Group gap="xs" wrap="nowrap">
-          <Box 
-            className={`${styles['state-circle']} ${styles[`state-${(lowerCase(state))}`]}`}
-            title={state}
-          />
-
-          {negativeKeyword === TARGET_STATES.ENABLED && (
-            <Box className={styles.negativeKeyword} title="Negative Keyword" >N</Box>
-          )}
-
-          {negativeTarget === TARGET_STATES.ENABLED && (
-            <Box className={styles.negativeTarget} title="Negative Target" >N</Box>
-          )}
-
-          <Text size="xs">{row.keyword}</Text>
-        </Group>
-      )
-    }
+    keyword: (row) => (
+      <KeywordColumn 
+        row={row}
+        state={keywords[row.keywordId]?.state || targets[row.keywordId]?.state}
+      />
+    ),
+    searchTerm: (row) => (
+      <SearchTermColumn 
+        row={row}
+        negativeKeywords={negativeKeywords}
+        negativeTargets={negativeTargets}
+      />
+    )
   }), [keywords, targets, negativeKeywords, negativeTargets])
 
   return (
